Add build sorting notes for Heart Beat and NSPC/Hect
Total number of known unique builds: 10 (8 for NSPC/Hect)

Heart Beat's VCMD set appears to be a partially offset version of Hect's NSPC
variant. They have been kept separate for now because more comparisons need to
be done, but Tsukada Tawada and Hitoshi Sakimoto are both credited in Dragon
Quest 3 and 6 for sound design as well as games prior to this one that use the
NSPC variant in question.