Add build sorting notes for Konami
Total number of known unique builds: Unknown (maximum is 35)

These notes are incomplete due to pointer differences AND the opening section of
code being overwritten after initialization of the sound driver for most builds.
However, there are a few cases for which I have confirmed a match with pointer
differences: there could be many more, but the game-specific data prior to this
may mean that dynamic compilation may also be involved, which would further
complicate the build sorting procedure.