Add build sorting notes for NSPC/Kazuo Sawa
Total number of known unique builds: 31 (10 for Taito, 21 for C-Lab)

These two are both under Kazuo Sawa because the music VCMDs are the same origin
wise. However, the SFX programming is different: Populous is the originator for
both of these, but the SFX programming carries over to games by Taito, as
opposed to a different set of builds that were used by C-Lab starting with Pro
Soccer.