Add build sorting notes for NSPC/Taro Bando
Total number of known unique builds: 9

There are multiple Gigaleak-era builds included in these sorting notes.
Specifically, they include a series of builds from Super Mario Kart, as well as
FX Trax. For Super Mario Kart specifically, both the ROM name and some extra
info is provided courtesy of research provided by the SMK Workshop Discord
Server to assist in identifying the sources for these particular cases.

This is the last build sorting note to be done for a variant or sound driver
that has a disassembly done. Beyond this point will gradually cover all of the
other sound drivers and N-SPC variants.

This commit closes #3.