Vibrato re-identified as Pitch Envelope for Compile
Vibrato usually assumes that the pitch moves back and forth from some sort of
fixed pitch... the way this is coded, though, means that the pitch does not have
to do that, and instead can go in whatever direction it desires. Thus, although
it can be used for vibrato, it has a larger variety of usages, and thus has been
re-identified as a pitch envelope.