New build found for Irem
R-Type III has a regional variant. The signature itself was changed because a
piece of text text identifying the version and build date had a one character
difference as well as an initialization difference in the DSP registers (the
pointer in question is after both of these past an opcode difference).