Corrections (and technically a new build) for NSPC/Intelligent Systems
It was discovered during the second pass that a bunch of starting locations were
recorded incorrectly. In addition, most of them had an initialization snippet of
code that was overwritten afterwards. A new build was technically found, but
only because it had the initialization snippet, whereas the other one didn't.