Add build sorting notes for David Whittaker
Total number of known unique builds: 31

This sound driver is documented by KungFuFurby on the SNESLab wiki, which means
that not counting the disassemblies for some fine print, it can be adapted to
vgmtrans as needed. This sound driver is notorious for having possibly the
largest number of unique builds versus the number of games it was made for, with
a triple pack of sound driver command featuresets, music and SFX code
differences along with a few games having multiple unique builds being
responsible for the high build count. The music and SFX VCMD code actually
almost never change between builds in the same game in general, they tend to
change between games.

One case is not counted as a unique build due to the difference only being in
memory location storage references, and that is between the two copies found in
the US version of Riddick Bowe Boxing.