Add build sorting notes for Sol/H-Sound
Total number of known unique builds: 3

This sound driver is documented by KungFuFurby on the SNESLab wiki, which means
that not counting the disassemblies for some fine print, it can be adapted to
vgmtrans as needed.

One case barely qualifies as a variant because it's technically an overwrite
followed by a branch of a single parameter, as well as a drumkit difference.