Marked down some identical build cases for Opus/SQ
Three of the builds matched a different version identifier build down to the
opcodes except for the version marker (or absence of one), which affected the
pointers.

In addition, Cacoma Knight also has an opcode-level match with a different
build. However, it is still counted as a unique build due to the only difference
being an array... which affects the parameter of two AND opcodes, which in turn
affects the music code, which means it must be treated separately for version
sorting reasons.