Added build sorting notes for Martin Walker
Total number of known unique builds: 11

There is a bonus section for an initialization build that is used in later
versions.

NOTE: Spawn (Psygnosis)'s build is known due to having been retrieved from David
Lowe's disk archives that were released. The game is otherwise not extant, and
is not relevant to the version that was released.